dataikos-backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.1

# Database
supabase==2.3.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP & File Handling
httpx==0.25.1
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.1.0

# Email
Jinja2==3.1.3

# Scheduler
pytz==2024.1

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1

# Production Server
gunicorn==21.2.0
